# test/openapi.json
{"openapi":"3.1.0","info":{"title":"EMCRM","description":"Event Management CRM System","version":"1.0.0"},"paths":{"/users/":{"post":{"tags":["Users"],"summary":"Create User","operationId":"create_user_users__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}":{"get":{"tags":["Users"],"summary":"Get User","operationId":"get_user_users__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Users"],"summary":"Update User","operationId":"update_user_users__user_id__put","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Users"],"summary":"Delete User","operationId":"delete_user_users__user_id__delete","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/events/":{"post":{"tags":["Events"],"summary":"Create Event","operationId":"create_event_events__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Event"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Event"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Events"],"summary":"Get Events","description":"Get all events with pagination","operationId":"get_events_events__get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Page number","default":1,"title":"Page"},"description":"Page number"},{"name":"size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Page size","default":50,"title":"Size"},"description":"Page size"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Page_Event_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/events/{event_id}":{"get":{"tags":["Events"],"summary":"Get Event","operationId":"get_event_events__event_id__get","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"string","title":"Event Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Event"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Events"],"summary":"Update Event","operationId":"update_event_events__event_id__put","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"string","title":"Event Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Event"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Event"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Events"],"summary":"Delete Event","operationId":"delete_event_events__event_id__delete","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"string","title":"Event Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/attend/":{"post":{"tags":["Attendance"],"summary":"Create Attendance","operationId":"create_attendance_attend__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventAttendance"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventAttendance"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/attend/user/{user_id}":{"get":{"tags":["Attendance"],"summary":"Get User Attendance","operationId":"get_user_attendance_attend_user__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Page number","default":1,"title":"Page"},"description":"Page number"},{"name":"size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Page size","default":50,"title":"Size"},"description":"Page size"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Page_EventAttendance_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/attend/event/{event_id}":{"get":{"tags":["Attendance"],"summary":"Get Event Attendance","operationId":"get_event_attendance_attend_event__event_id__get","parameters":[{"name":"event_id","in":"path","required":true,"schema":{"type":"string","title":"Event Id"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Page number","default":1,"title":"Page"},"description":"Page number"},{"name":"size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Page size","default":50,"title":"Size"},"description":"Page size"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Page_EventAttendance_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/search/query_users":{"post":{"tags":["Search"],"summary":"Filter Users Opensearch","operationId":"filter_users_opensearch_search_query_users_post","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Page"}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Size"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserFilter"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSearchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/email/":{"get":{"tags":["Email"],"summary":"Get Email Requests","operationId":"get_email_requests_email__get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Page number","default":1,"title":"Page"},"description":"Page number"},{"name":"size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Page size","default":50,"title":"Size"},"description":"Page size"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Page_EmailRequest_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/email/send_emails":{"post":{"tags":["Email"],"summary":"Send Email To Filtered Users","operationId":"send_email_to_filtered_users_email_send_emails_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailRequest"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/email/{email_id}":{"post":{"tags":["Email"],"summary":"Get Email Sent Request","operationId":"get_email_sent_request_email__email_id__post","parameters":[{"name":"email_id","in":"path","required":true,"schema":{"type":"string","title":"Email Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailRequest"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/email/sent/{email_id}":{"post":{"tags":["Email"],"summary":"Get Emails Status","operationId":"get_emails_status_email_sent__email_id__post","parameters":[{"name":"email_id","in":"path","required":true,"schema":{"type":"string","title":"Email Id"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Page number","default":1,"title":"Page"},"description":"Page number"},{"name":"size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Page size","default":50,"title":"Size"},"description":"Page size"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Page_Email_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/email/user/{user_id}":{"post":{"tags":["Email"],"summary":"Get Emails Status","operationId":"get_emails_status_email_user__user_id__post","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Page number","default":1,"title":"Page"},"description":"Page number"},{"name":"size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Page size","default":50,"title":"Size"},"description":"Page size"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Page_Email_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Check system health","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Email":{"properties":{"email_id":{"type":"string","maxLength":50,"title":"Email Id"},"user_id":{"type":"string","maxLength":50,"title":"User Id"},"status":{"anyOf":[{"$ref":"#/components/schemas/EmailStatusEnum"},{"type":"null"}]},"createdAt":{"type":"string","format":"date-time","title":"Createdat"}},"type":"object","required":["email_id","user_id"],"title":"Email"},"EmailRequest":{"properties":{"email_id":{"type":"string","title":"Email Id"},"filter":{"$ref":"#/components/schemas/UserFilter"},"createdAt":{"type":"string","format":"date-time","title":"Createdat"},"totalRecipients":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Totalrecipients","default":0},"status":{"anyOf":[{"$ref":"#/components/schemas/EmailStatusEnum"},{"type":"null"}],"default":"sent"},"subject":{"type":"string","maxLength":50,"title":"Subject"},"body":{"type":"string","maxLength":1000,"title":"Body"}},"type":"object","required":["filter","subject","body"],"title":"EmailRequest"},"EmailStatusEnum":{"type":"string","enum":["error","pending","sending","sent"],"title":"EmailStatusEnum"},"Event":{"properties":{"id":{"type":"string","title":"Id"},"slug":{"type":"string","maxLength":50,"minLength":3,"pattern":"^[a-z0-9]+(?:-[a-z0-9]+)*$","title":"Slug"},"title":{"type":"string","maxLength":100,"minLength":3,"title":"Title"},"description":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Description"},"startAt":{"type":"string","format":"date-time","title":"Startat"},"endAt":{"type":"string","format":"date-time","title":"Endat"},"venue":{"anyOf":[{"type":"string","maxLength":200},{"type":"null"}],"title":"Venue"},"maxCapacity":{"anyOf":[{"type":"integer","minimum":1.0},{"type":"null"}],"title":"Maxcapacity"},"owner":{"type":"string","title":"Owner"},"hosts":{"items":{"type":"string"},"type":"array","title":"Hosts","default":[]}},"type":"object","required":["slug","title","startAt","endAt","owner"],"title":"Event"},"EventAttendance":{"properties":{"user_id":{"type":"string","maxLength":50,"title":"User Id"},"event_id":{"type":"string","maxLength":50,"title":"Event Id"},"attended":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Attended","default":"yes"},"createdAt":{"type":"string","format":"date-time","title":"Createdat"}},"type":"object","required":["user_id","event_id"],"title":"EventAttendance"},"GenderEnum":{"type":"string","enum":["male","female","other"],"title":"GenderEnum"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Page_EmailRequest_":{"properties":{"items":{"items":{"$ref":"#/components/schemas/EmailRequest"},"type":"array","title":"Items"},"total":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Total"},"page":{"anyOf":[{"type":"integer","minimum":1.0},{"type":"null"}],"title":"Page"},"size":{"anyOf":[{"type":"integer","minimum":1.0},{"type":"null"}],"title":"Size"},"pages":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Pages"}},"type":"object","required":["items","page","size"],"title":"Page[EmailRequest]"},"Page_Email_":{"properties":{"items":{"items":{"$ref":"#/components/schemas/Email"},"type":"array","title":"Items"},"total":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Total"},"page":{"anyOf":[{"type":"integer","minimum":1.0},{"type":"null"}],"title":"Page"},"size":{"anyOf":[{"type":"integer","minimum":1.0},{"type":"null"}],"title":"Size"},"pages":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Pages"}},"type":"object","required":["items","page","size"],"title":"Page[Email]"},"Page_EventAttendance_":{"properties":{"items":{"items":{"$ref":"#/components/schemas/EventAttendance"},"type":"array","title":"Items"},"total":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Total"},"page":{"anyOf":[{"type":"integer","minimum":1.0},{"type":"null"}],"title":"Page"},"size":{"anyOf":[{"type":"integer","minimum":1.0},{"type":"null"}],"title":"Size"},"pages":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Pages"}},"type":"object","required":["items","page","size"],"title":"Page[EventAttendance]"},"Page_Event_":{"properties":{"items":{"items":{"$ref":"#/components/schemas/Event"},"type":"array","title":"Items"},"total":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Total"},"page":{"anyOf":[{"type":"integer","minimum":1.0},{"type":"null"}],"title":"Page"},"size":{"anyOf":[{"type":"integer","minimum":1.0},{"type":"null"}],"title":"Size"},"pages":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Pages"}},"type":"object","required":["items","page","size"],"title":"Page[Event]"},"User":{"properties":{"id":{"type":"string","title":"Id"},"firstName":{"type":"string","maxLength":50,"title":"Firstname"},"lastName":{"type":"string","maxLength":50,"title":"Lastname"},"email":{"type":"string","format":"email","title":"Email"},"phoneNumber":{"anyOf":[{"type":"string","maxLength":20,"minLength":7},{"type":"null"}],"title":"Phonenumber"},"avatar":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Avatar"},"gender":{"anyOf":[{"$ref":"#/components/schemas/GenderEnum"},{"type":"null"}]},"jobTitle":{"anyOf":[{"type":"string","maxLength":50},{"type":"null"}],"title":"Jobtitle"},"company":{"anyOf":[{"type":"string","maxLength":50},{"type":"null"}],"title":"Company"},"city":{"anyOf":[{"type":"string","maxLength":50},{"type":"null"}],"title":"City"},"state":{"anyOf":[{"type":"string","maxLength":50},{"type":"null"}],"title":"State"},"attendedCount":{"type":"integer","minimum":0.0,"title":"Attendedcount","default":0},"hostedCount":{"type":"integer","minimum":0.0,"title":"Hostedcount","default":0}},"type":"object","required":["firstName","lastName","email"],"title":"User"},"UserFilter":{"properties":{"company":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Company"},"jobTitle":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Jobtitle"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"},"minAttended":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Minattended"},"maxAttended":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Maxattended"},"minHosted":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Minhosted"},"maxHosted":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Maxhosted"}},"type":"object","title":"UserFilter"},"UserSearchResponse":{"properties":{"total":{"type":"integer","title":"Total"},"users":{"items":{"$ref":"#/components/schemas/User"},"type":"array","title":"Users"}},"type":"object","required":["total","users"],"title":"UserSearchResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}